move views.chant.ChantListView -> views.source.SourceBrowseChantsView
